Ensure expand_operator output dtype

+++ src/qutip_qip/operations/gates.py
@@ -1326,7 +1326,7 @@ def expand_operator(
     for i, ind in enumerate(rest_pos):
         new_order[ind] = rest_qubits[i]
     id_list = [identity(dims[i]) for i in rest_pos]
-    return tensor([oper] + id_list).permute(new_order)
+    return tensor([oper] + id_list).permute(new_order).to(dtype)
 
 
 def gate_sequence_product(
